% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapboost_estimator.R
\docType{class}
\name{SHAPBoostEstimator-class}
\alias{SHAPBoostEstimator-class}
\alias{SHAPBoostEstimator}
\title{SHAPBoostEstimator Class}
\description{
This class implements the SHAPBoost algorithm for feature selection and model fitting.
It is designed to be extended by specific implementations such as SHAPBoostRegressor and
SHAPBoostSurvival.
}
\section{Fields}{

\describe{
\item{\code{estimators}}{A list of estimators used in the SHAPBoost algorithm.}

\item{\code{loss}}{A function representing the loss function used in the algorithm.}

\item{\code{metric}}{A character string representing the evaluation metric.}

\item{\code{number_of_folds}}{The number of folds for cross-validation.}

\item{\code{epsilon}}{A small value to determine convergence.}

\item{\code{max_number_of_features}}{The maximum number of features to select.}

\item{\code{siso_ranking_size}}{The number of features to consider in the SISO ranking.}

\item{\code{siso_order}}{The order of combinations to consider in SISO.}

\item{\code{reset}}{A logical indicating whether to reset the weights.}

\item{\code{xgb_importance}}{The method for calculating feature importance in XGBoost.}

\item{\code{num_resets}}{The number of resets allowed.}

\item{\code{fold_random_state}}{The random state for reproducibility in cross-validation.}

\item{\code{verbose}}{The verbosity level of the output.}

\item{\code{stratification}}{A logical indicating whether to use stratified sampling.}

\item{\code{use_shap}}{A logical indicating whether to use SHAP values for feature importance.}

\item{\code{collinearity_check}}{A logical indicating whether to check for collinearity.}

\item{\code{correlation_threshold}}{The threshold for correlation to consider features as collinear.}
}}


